Create a separate package for Protobuf Gradle plugin integration.
