+++ plugin/src/main/java/io/spine/tools/bootstrap/BootstrapPlugin.java
@@ -20,6 +20,7 @@
 
 package io.spine.tools.bootstrap;
 
+import io.spine.tools.bootstrap.protobuf.ProtobufGenerator;
 import io.spine.tools.gradle.SpinePlugin;
 import org.gradle.api.Project;
